+++ tests/FeatureTests/Endpoint/User/PostChangePasswordTest.php
@@ -46,7 +46,7 @@ class PostChangePasswordTest extends BaseRequestTestCase
                 'password' => self::NEW_PASSWORD,
             ]
         );
-        $this->assertSame(200, $createTokenWithNewPasswordResponse->getStatusCode());
+        $this->assertSame(201, $createTokenWithNewPasswordResponse->getStatusCode());
 
         $changePasswordWithOldCredentialsFails = $this->runPostRequest(
             '/change-password',