+++ tests/FeatureTests/Endpoint/User/PostTokenTest.php
@@ -27,7 +27,7 @@ class PostTokenTest extends BaseRequestTestCase
             ]
         );
 
-        $this->assertSame(200, $response->getStatusCode());
+        $this->assertSame(201, $response->getStatusCode());
         $this->assertSame('application/json; charset=utf-8', $response->getHeader('content-type')[0]);
         $body = \Safe\json_decode((string) $response->getBody(), true);
         $this->assertSame('_TokenResponse', $body['type']);
@@ -57,7 +57,7 @@ class PostTokenTest extends BaseRequestTestCase
             ]
         );
 
-        $this->assertSame(200, $response->getStatusCode());
+        $this->assertSame(201, $response->getStatusCode());
         $this->assertSame('application/json; charset=utf-8', $response->getHeader('content-type')[0]);
         $body = \Safe\json_decode((string) $response->getBody(), true);
         $this->assertSame('_TokenResponse', $body['type']);
@@ -87,7 +87,7 @@ class PostTokenTest extends BaseRequestTestCase
             ]
         );
 
-        $this->assertSame(200, $response->getStatusCode());
+        $this->assertSame(201, $response->getStatusCode());
         $this->assertSame('application/json; charset=utf-8', $response->getHeader('content-type')[0]);
         $body = \Safe\json_decode((string) $response->getBody(), true);
         $this->assertSame('_TokenResponse', $body['type']);
@@ -117,7 +117,7 @@ class PostTokenTest extends BaseRequestTestCase
             ]
         );
 
-        $this->assertSame(200, $response->getStatusCode());
+        $this->assertSame(201, $response->getStatusCode());
         $this->assertSame('application/json; charset=utf-8', $response->getHeader('content-type')[0]);
         $body = \Safe\json_decode((string) $response->getBody(), true);
         $this->assertSame('_TokenResponse', $body['type']);