+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Upgrade to Alpine 3.22, closes #389.
 - Upgrade PHP to 8.4.11, closes #395.
+### Fixed
+- Post token endpoint returns status code 201 instead of 200, closes #392.
 
 ## 0.1.22 - 2025-05-21
 ### Changed